requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-core
openai
